@@ -3,7 +3,7 @@ description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 0.0.1+1
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: ">=2.18.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -30,19 +30,19 @@ dependencies:
   cupertino_icons: ^1.0.2
   equatable: ^2.0.3
   expandable: ^5.0.1
-  firebase_auth: ^3.3.11
-  firebase_core: ^1.13.1
-  firebase_database: ^9.0.15
+  firebase_auth: ^4.14.1
+  firebase_core: ^2.23.0
+  firebase_database: ^10.3.5
   flutter:
     sdk: flutter
   flutter_bloc: ^8.0.1
-  flutter_html: ^2.2.1
+  flutter_html: 3.0.0-beta.2
   flutter_inappwebview: ^5.3.2
-  google_fonts: 2.1.1
+  google_fonts: ^6.1.0
   url_launcher: ^6.0.20
 
 dev_dependencies:
-  flutter_lints: ^1.0.0
+  flutter_lints: ^3.0.1
   flutter_test:
     sdk: flutter
 
