@@ -1,7 +1,6 @@
-import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:ta_bsi/src/data/dataSources/remote/auth_user_service.dart';
-import 'package:ta_bsi/src/data/dataSources/remote/user_service.dart';
 import 'package:ta_bsi/src/data/models/user_model.dart';
 
 part 'auth_state.dart';
