@@ -1,5 +1,5 @@
-import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:ta_bsi/src/data/dataSources/remote/module_service.dart';
 import 'package:ta_bsi/src/data/models/module_model.dart';
 
