@@ -1,4 +1,4 @@
-import 'package:bloc/bloc.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 
 class PageCubit extends Cubit<int> {
   PageCubit() : super(0);
