@@ -1,4 +1,4 @@
-import 'package:bloc/bloc.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:equatable/equatable.dart';
 import 'package:ta_bsi/src/data/dataSources/remote/quiz_service.dart';
 import 'package:ta_bsi/src/data/models/quiz_model.dart';
@@ -9,24 +9,30 @@ class QuizCubit extends Cubit<QuizState> {
   QuizCubit() : super(QuizInitial());
 
   /// variabel list Quiz
-  List<QuizModel> quiz = [];
+  List<QuizModel> _quiz = [];
+  String typeModule = '';
+  String idQuiz = '';
 
   /// fungsi setQuiz untuk mengisi variabel Quiz
   void setQuiz(List<QuizModel> newQuiz) {
-    quiz = newQuiz;
+    _quiz = newQuiz;
   }
 
-  void fetchListQuiz(String module) async {
+  void setTypeModule(String newValue) {
+    typeModule = newValue;
+  }
+
+  void setIdQuiz(String newValue) {
+    idQuiz = newValue;
+  }
+
+  void fetchListQuiz() async {
     try {
       emit(QuizLoading());
 
-      /// jika [Quiz] masih kosong paggil fungsi setQuiz dengan
-      /// paramater QuizService.fetchQuiz()
-      if (quiz.isEmpty) {
-        setQuiz(await QuizService.fetchQuiz(module));
-      }
+      setQuiz(await QuizService.fetchQuiz(typeModule));
 
-      emit(QuizSuccess(quiz));
+      emit(QuizSuccess(_quiz));
     } catch (e) {
       emit(QuizFailed(e.toString()));
       print('error quiz cubit $e');
